refacto task state code into its own file
